+++ examples/mugen/retrieval/definitions.py
@@ -7,7 +7,7 @@
 from dataclasses import dataclass
 from typing import Any, Dict, Optional, Tuple
 
-from examples.mugen.data.audio_utils import AUDIO_SAMPLE_LENGTH, AUDIO_SAMPLE_RATE
+from examples.mugen.data.mugen_dataset import MUGENDatasetArgs
 
 from torchmultimodal.transforms.video_transform import (
     DEFAULT_MEAN,
@@ -29,39 +29,9 @@ class BertTextTransformArgs:
 @dataclass
 class VideoTransformArgs:
     time_samples: int = MUGEN_DEFAULT_TIME_SAMPLES
-    mean: Tuple[float] = tuple(DEFAULT_MEAN)
-    std: Tuple[float] = tuple(DEFAULT_STD)
+    mean: Tuple[float] = DEFAULT_MEAN
+    std: Tuple[float] = DEFAULT_STD
     resize_shape: Tuple[int, int] = DEFAULT_RESIZE_SHAPE
-
-
-@dataclass
-class MUGENDatasetArgs:
-    data_path: str = "datasets/coinrun/coinrun_dataset_jsons/release"
-    asset_path: str = "datasets/coinrun/assets"
-    sample_every_n_frames: int = 3
-    sequence_length: int = 32
-    resolution: int = 256
-    audio_sample_rate: int = AUDIO_SAMPLE_RATE
-    audio_sample_length: int = AUDIO_SAMPLE_LENGTH
-    bbox_smap_for_agent: bool = (
-        True  # render smap for mugen (and shield) as bounding boxes
-    )
-    bbox_smap_for_monsters: bool = True  # render smap for monsters as bounding boxes
-    use_manual_annotation: bool = False  # if True will only use videos with manual annotation and skip those without
-    use_auto_annotation: bool = (
-        True  # if True will only use videos with auto annotation and skip those without
-    )
-    use_downsampled_trainset: bool = (
-        False  # if True will only use downsampled training set
-    )
-    fixed_start_idx: bool = True  # fx starting game frame idx to 0
-    get_game_frame: bool = False  # load video data
-    get_seg_map: bool = False  # load semantic map
-    get_text_desc: bool = False  # load text data
-    get_audio: bool = (
-        False  # load full mix audio for each video, for audio generation models
-    )
-    debug: bool = False
 
 
 @dataclass
